this helps?
git-svn-id: file:///svn/rdup/trunk@1382 0f4cb41f-7f08-0410-98f9-8ec83d93c06f

+++ gfunc.c
@@ -324,10 +324,9 @@ gfunc_write(gpointer data, __attribute__((unused)) gpointer value, gpointer fp)
 	if (value == NO_PRINT)
 		return FALSE;
 
-	/* mode_path */
 	/* this is used to create our filelist */
-	fprintf((FILE*) fp, "%d %d %d %d %s", (int)e->f_mode, (int)e->f_dev, 
-			(int)e->f_ino, (int)e->f_name_size, e->f_name);
+	fprintf((FILE*) fp, "%ld %ld %ld %ld %s", (long int)e->f_mode, (long int)e->f_dev, 
+			(long int)e->f_ino, (long int)e->f_name_size, e->f_name);
 	if (opt_null) {
 		fputc('\0', (FILE*)fp);
 	} else {
